provision: Split up provision.py and provision_hex.cmake
Split up provision.py and provision_hex.cmake into bl_storage and
counters variants as they are unrelated.
