+++ samples/nrf5340/netboot/Kconfig
@@ -12,14 +12,13 @@ config NETBOOT_MIN_PARTITION_SIZE
 config B0N_SIZE
 	hex
 	prompt "Size of the B0n partition."
-	default 0x2580 if PM_PARTITION_SIZE_PROVISION=0x280 && NETBOOT_MIN_PARTITION_SIZE
-	default 0x8580 if PM_PARTITION_SIZE_PROVISION=0x280
-	default 0x8800
+	default 0x2510 if NETBOOT_MIN_PARTITION_SIZE
+	default 0x8510
 	help
 		Since the locking granularity is 2kB, ensure that B0N_SIZE is
 		set such that the size of the `b0n_container` is aligned to this
 		size. If the minimal config is enabled the binary will be around
-		5kB. Note: The provision partition is 0x280 bytes.
+		5kB. Note: The provision partition is 0x2F0 bytes.
 
 endmenu
