+++ subsys/bootloader/include/dummy_values/pm_config.h
@@ -11,8 +11,8 @@
 #define PM_B0N_SIZE 0x1000
 #define PM_APP_ADDRESS 0x7000
 #define PM_APP_SIZE 0x1000
-#define PM_PROVISION_ADDRESS 0x7000
-#define PM_PROVISION_SIZE 0x1000
+#define PM_PROVISION_BL_STORAGE_ADDRESS 0x7000
+#define PM_PROVISION_BL_STORAGE_SIZE 0x1000
 #define PM_S0_SIZE 0x10000
 #define PM_S1_SIZE 0x10000
 #define PM__PCD_SRAM_ADDRESS 0x2006e000
